quality_metrics/requirements.txt
amqp==5.3.1
billiard==4.2.1
boto3==1.37.33
botocore==1.37.33
celery==5.5.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
imageio==2.37.0
jmespath==1.0.1
kombu==5.5.2
lazy_loader==0.4
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pyav==14.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.2.1
s3transfer==0.11.4
scikit-image==0.25.2
scipy==1.15.2
six==1.17.0
SQLAlchemy==2.0.40
tifffile==2025.3.30
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
